Refactored existing functionality
Added an optional url property to beer model which is added in
beer.service.  This url is the url of the beer detail page.  Now
templates don't need to know how to reach the beer detail page.

BeerService and BeerListComponent were refactored to use BeerService.getBeers() for searching.  In the process of refactoring BeerService, basic query building functionality was added.