Added srm attribute to beer model.
The srm of a beer can be used to calculate the color of a beer.  By
adding it to the model it may be possible to search by beer color
and find beers of a similar srm (and color).

See also: fb47b3dad44d76bfd52dae98f14608f2e8632959